fix: move PasswordModal to beta folder
