+++ web/src/classic/components/molecules/Settings/Project/Plugin/PluginSection/PluginInstall/PublicRepo/index.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 
+import TextBox from "@reearth/beta/components/TextBox";
 import Box from "@reearth/classic/components/atoms/Box";
 import Button from "@reearth/classic/components/atoms/Button";
 import { Icons } from "@reearth/classic/components/atoms/Icon";
 import Loading from "@reearth/classic/components/atoms/Loading";
 import Modal from "@reearth/classic/components/atoms/Modal";
 import Text from "@reearth/classic/components/atoms/Text";
-import TextBox from "@reearth/classic/components/atoms/TextBox";
 import { useT } from "@reearth/services/i18n";
 import { useTheme } from "@reearth/services/theme";
