+++ web/src/classic/components/molecules/Settings/Workspace/AddMemberModal/index.tsx
@@ -1,10 +1,10 @@
 import React, { useCallback, useState, useEffect } from "react";
 
+import TextBox from "@reearth/beta/components/TextBox";
 import Button from "@reearth/classic/components/atoms/Button";
 import Icon from "@reearth/classic/components/atoms/Icon";
 import Modal from "@reearth/classic/components/atoms/Modal";
 import Text from "@reearth/classic/components/atoms/Text";
-import TextBox from "@reearth/classic/components/atoms/TextBox";
 import { useT } from "@reearth/services/i18n";
 import { styled, useTheme } from "@reearth/services/theme";
