+++ web/src/classic/theme/reearthTheme/common/index.ts
@@ -1,8 +1,8 @@
 import colors, { Colors } from "./colors";
-import { MetricsSizesType, metricsSizes } from "./metrics";
+import { MetricsSizesType, metricsSizes } from "../../../../beta/utils/metrics";
 import zIndexes, { ZIndex } from "./zIndex";
 
-export type { MetricsSizes } from "./metrics";
+export type { MetricsSizes } from "../../../../beta/utils/metrics";
 
 type Common = {
   zIndexes: ZIndex;
